requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==4.2.13
django-summernote==0.8.20.0
gunicorn==20.1.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
